Shakespeare-generating character-level RNN.

+++ vel/util/tensor_util.py
@@ -3,7 +3,7 @@ import torch
 
 def one_hot_encoding(input_tensor, num_labels):
     """ One-hot encode labels from input """
-    return torch.eye(num_labels, device=input_tensor.device)[input_tensor]
+    return torch.eye(num_labels, device=input_tensor.device)[input_tensor.to(torch.long)]
 
 
 def merge_first_two_dims(tensor):
